// src/app/api/model/carryoverApiModel.ts
/**
 * Evolveum Day Off Planner API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface CarryoverApiModel { 
    readonly user: string;
    readonly leaveType: string;
    readonly year: number;
    readonly carryover: number;
}